updated FAQ, fixed some type conversions
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@4312 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/ANALYSIS/ID/HiddenMarkovModel.C
@@ -340,7 +340,7 @@ namespace OpenMS
 	
 	UInt HiddenMarkovModel::getNumberOfStates() const
 	{
-		return states_.size();
+		return (UInt)states_.size();
 	}
 
 	void HiddenMarkovModel::addNewState(HMMState* s)
